--- living-quantum-monitor.py
@@ -1,384 +1,0 @@
-# modules/quantum_cea/living_quantum_monitor.py
-"""
-Living Quantum CEA Monitor - Phase 1 Implementation
-Evidence-led monitoring and mapping for biological quantum computing
-
-Following OSCE Recommendation Framework:
-- Scientific rigor through measurement
-- Ethical transparency through open data
-- System health as primary concern
-"""
-
-import asyncio
-import numpy as np
-from typing import Dict, List, Optional, Tuple
-from datetime import datetime, timedelta
-from dataclasses import dataclass, field
-from enum import Enum
-import json
-
-from osce.core.base import OSCEModule
-from osce.utils.logging import get_logger
-
-logger = get_logger(__name__)
-
-class MonitoringPhase(Enum):
-    """Staged progression following recommendation framework"""
-    BASELINE = "baseline"  # Map normal ecosystem function
-    COHERENCE = "coherence"  # Detect natural synchronization
-    CORRELATION = "correlation"  # Find quantum-like behaviors
-    EXPERIMENTAL = "experimental"  # Careful quantum tests
-
-@dataclass
-class EcosystemHealth:
-    """Primary concern: ecosystem wellbeing"""
-    timestamp: datetime
-    overall_health: float  # 0-1 score
-    plant_vitality: Dict[str, float]
-    fungal_network_integrity: float
-    beneficial_organism_count: Dict[str, int]
-    soil_health_metrics: Dict[str, float]
-    stress_indicators: List[str]
-    recommendations: List[str]
-
-@dataclass
-class CoherenceObservation:
-    """Document potential coherence with uncertainty"""
-    timestamp: datetime
-    observation_type: str
-    measured_value: float
-    expected_value: float
-    confidence: float  # 0-1, documenting uncertainty
-    interpretation: str
-    raw_data_reference: str  # For transparency
-
-@dataclass
-class EthicalCheckpoint:
-    """Ensure consent and review at each stage"""
-    phase: MonitoringPhase
-    criteria_met: Dict[str, bool]
-    community_input: List[str]
-    ethical_concerns: List[str]
-    approval_status: bool
-    reviewers: List[str]
-
-class LivingQuantumMonitor(OSCEModule):
-    """
-    Phase 1: Evidence-led monitoring and baseline establishment
-    
-    Following principles:
-    - System health over experimental novelty
-    - Radical transparency in all measurements
-    - Community consent before progression
-    - Document uncertainty honestly
-    """
-    
-    def __init__(self, config: Dict):
-        super().__init__(config)
-        
-        # Current monitoring phase
-        self.current_phase = MonitoringPhase.BASELINE
-        
-        # Health monitoring (priority #1)
-        self.health_threshold = config.get('min_health_score', 0.8)
-        self.health_history = []
-        
-        # Coherence detection
-        self.coherence_observations = []
-        self.baseline_measurements = {}
-        
-        # Ethical governance
-        self.ethical_checkpoints = []
-        self.community_reviewers = config.get('community_reviewers', [])
-        
-        # Data transparency
-        self.public_data_path = config.get('public_data_path', 'data/public/')
-        self.anomaly_log = []
-        
-    async def initialize(self):
-        """Initialize monitoring with health checks first"""
-        logger.info("Initializing Living Quantum CEA Monitor - Phase 1")
-        
-        # Verify ecosystem health before any experiments
-        health = await self._assess_ecosystem_health()
-        if health.overall_health < self.health_threshold:
-            logger.warning(f"Ecosystem health below threshold: {health.overall_health}")
-            await self._initiate_health_recovery(health)
-            return
-            
-        # Start baseline monitoring
-        asyncio.create_task(self._baseline_monitoring_loop())
-        
-        # Initialize transparent data sharing
-        await self._setup_public_data_sharing()
-        
-        logger.info("Phase 1 monitoring initialized - prioritizing system health")
-        
-    async def _baseline_monitoring_loop(self):
-        """Establish baseline before any intervention"""
-        while self.current_phase == MonitoringPhase.BASELINE:
-            try:
-                # Health check first (always)
-                health = await self._assess_ecosystem_health()
-                self.health_history.append(health)
-                
-                if health.overall_health < self.health_threshold:
-                    logger.warning("Health below threshold - pausing experiments")
-                    await self._pause_all_experiments()
-                    
-                # Baseline measurements
-                measurements = await self._collect_baseline_measurements()
-                self.baseline_measurements[datetime.utcnow()] = measurements
-                
-                # Look for natural coherence (no intervention)
-                natural_coherence = await self._detect_natural_coherence(measurements)
-                if natural_coherence:
-                    self.coherence_observations.append(natural_coherence)
-                    
-                # Transparent data publication
-                await self._publish_measurements(measurements, health)
-                
-            except Exception as e:
-                logger.error(f"Baseline monitoring error: {e}")
-                self.anomaly_log.append({
-                    'timestamp': datetime.utcnow(),
-                    'error': str(e),
-                    'phase': self.current_phase.value
-                })
-                
-            await asyncio.sleep(300)  # 5-minute intervals
-            
-    async def _assess_ecosystem_health(self) -> EcosystemHealth:
-        """Comprehensive health assessment - top priority"""
-        # Plant health
-        plants = await self.env.get_all_plants()
-        plant_vitality = {}
-        for plant in plants:
-            vitality = await self._assess_plant_vitality(plant['id'])
-            plant_vitality[plant['id']] = vitality
-            
-        # Fungal network health
-        fungal_integrity = await self._assess_fungal_network()
-        
-        # Beneficial organisms
-        organism_counts = await self._count_beneficial_organisms()
-        
-        # Soil health
-        soil_metrics = await self._analyze_soil_health()
-        
-        # Calculate overall health
-        overall = np.mean([
-            np.mean(list(plant_vitality.values())),
-            fungal_integrity,
-            self._score_organism_balance(organism_counts),
-            np.mean(list(soil_metrics.values()))
-        ])
-        
-        # Identify stress indicators
-        stress_indicators = []
-        if overall < 0.8:
-            stress_indicators.extend(self._identify_stressors(
-                plant_vitality, fungal_integrity, organism_counts, soil_metrics
-            ))
-            
-        # Generate recommendations
-        recommendations = self._generate_health_recommendations(
-            stress_indicators, overall
-        )
-        
-        return EcosystemHealth(
-            timestamp=datetime.utcnow(),
-            overall_health=float(overall),
-            plant_vitality=plant_vitality,
-            fungal_network_integrity=float(fungal_integrity),
-            beneficial_organism_count=organism_counts,
-            soil_health_metrics=soil_metrics,
-            stress_indicators=stress_indicators,
-            recommendations=recommendations
-        )
-        
-    async def _detect_natural_coherence(self, measurements: Dict) -> Optional[CoherenceObservation]:
-        """Detect coherence without intervention - respecting natural patterns"""
-        # Look for synchronization in bioelectric signals
-        if 'bioelectric_signals' in measurements:
-            signals = measurements['bioelectric_signals']
-            
-            # Calculate phase coherence
-            coherence = await self._calculate_phase_coherence(signals)
-            
-            if coherence > 0.7:  # Significant natural coherence
-                return CoherenceObservation(
-                    timestamp=datetime.utcnow(),
-                    observation_type='natural_bioelectric_coherence',
-                    measured_value=coherence,
-                    expected_value=0.3,  # Random expectation
-                    confidence=0.8,  # We're still learning
-                    interpretation="Natural synchronization detected - no intervention",
-                    raw_data_reference=f"{self.public_data_path}/coherence_{datetime.utcnow().isoformat()}.json"
-                )
-                
-        return None
-        
-    async def request_phase_advancement(self, next_phase: MonitoringPhase) -> bool:
-        """
-        Request to advance to next phase with full review
-        Following staged gate review process
-        """
-        logger.info(f"Requesting advancement to {next_phase.value}")
-        
-        # Check technical readiness
-        technical_ready = await self._check_technical_readiness(next_phase)
-        
-        # Check ethical safety
-        ethical_ready = await self._check_ethical_safety(next_phase)
-        
-        # Get community buy-in
-        community_ready = await self._get_community_approval(next_phase)
-        
-        # Create checkpoint record
-        checkpoint = EthicalCheckpoint(
-            phase=next_phase,
-            criteria_met={
-                'technical': technical_ready,
-                'ethical': ethical_ready,
-                'community': community_ready,
-                'health': await self._check_system_health()
-            },
-            community_input=await self._gather_community_input(),
-            ethical_concerns=await self._identify_ethical_concerns(next_phase),
-            approval_status=all([technical_ready, ethical_ready, community_ready]),
-            reviewers=self.community_reviewers
-        )
-        
-        self.ethical_checkpoints.append(checkpoint)
-        
-        # Log decision transparently
-        await self._publish_phase_decision(checkpoint)
-        
-        if checkpoint.approval_status:
-            self.current_phase = next_phase
-            logger.info(f"Advanced to phase: {next_phase.value}")
-            await self._initialize_phase(next_phase)
-            return True
-        else:
-            logger.info(f"Phase advancement denied. Concerns: {checkpoint.ethical_concerns}")
-            return False
-            
-    async def _publish_measurements(self, measurements: Dict, health: EcosystemHealth):
-        """Radical transparency - publish all data"""
-        public_data = {
-            'timestamp': datetime.utcnow().isoformat(),
-            'phase': self.current_phase.value,
-            'health': {
-                'overall': health.overall_health,
-                'stress_indicators': health.stress_indicators,
-                'recommendations': health.recommendations
-            },
-            'measurements': self._sanitize_measurements(measurements),
-            'coherence_observations': [
-                {
-                    'timestamp': obs.timestamp.isoformat(),
-                    'type': obs.observation_type,
-                    'value': obs.measured_value,
-                    'confidence': obs.confidence,
-                    'interpretation': obs.interpretation
-                }
-                for obs in self.coherence_observations[-10:]  # Last 10
-            ],
-            'anomalies': self.anomaly_log[-5:],  # Last 5
-            'next_review': (datetime.utcnow() + timedelta(days=7)).isoformat()
-        }
-        
-        # Write to public data path
-        filename = f"{self.public_data_path}/monitor_{datetime.utcnow().strftime('%Y%m%d_%H%M%S')}.json"
-        with open(filename, 'w') as f:
-            json.dump(public_data, f, indent=2)
-            
-        # Also publish summary to dashboard
-        await self.env.publish_dashboard_data('quantum_cea_monitor', public_data)
-        
-    async def _check_technical_readiness(self, phase: MonitoringPhase) -> bool:
-        """Verify technical prerequisites for phase"""
-        if phase == MonitoringPhase.COHERENCE:
-            # Need stable baseline
-            return len(self.baseline_measurements) >= 100
-            
-        elif phase == MonitoringPhase.CORRELATION:
-            # Need observed natural coherence
-            return len(self.coherence_observations) >= 10
-            
-        elif phase == MonitoringPhase.EXPERIMENTAL:
-            # Need correlation evidence
-            return await self._verify_correlation_evidence()
-            
-        return False
-        
-    async def _check_ethical_safety(self, phase: MonitoringPhase) -> bool:
-        """Ensure ethical considerations are addressed"""
-        concerns = await self._identify_ethical_concerns(phase)
-        
-        if concerns:
-            # Address each concern
-            for concern in concerns:
-                resolution = await self._propose_resolution(concern)
-                if not resolution:
-                    return False
-                    
-        return True
-        
-    async def _get_community_approval(self, phase: MonitoringPhase) -> bool:
-        """Community consultation and approval"""
-        # Notify reviewers
-        proposal = await self._create_phase_proposal(phase)
-        
-        approvals = []
-        for reviewer in self.community_reviewers:
-            try:
-                approval = await self._request_reviewer_approval(reviewer, proposal)
-                approvals.append(approval)
-            except Exception as e:
-                logger.error(f"Failed to get approval from {reviewer}: {e}")
-                approvals.append(False)
-                
-        # Require majority approval
-        return sum(approvals) > len(approvals) / 2
-        
-    def _generate_health_recommendations(self, 
-                                       stress_indicators: List[str], 
-                                       overall_health: float) -> List[str]:
-        """Generate actionable health recommendations"""
-        recommendations = []
-        
-        if overall_health < 0.6:
-            recommendations.append("URGENT: Pause all experiments and focus on recovery")
-            
-        for indicator in stress_indicators:
-            if 'water' in indicator.lower():
-                recommendations.append("Adjust irrigation schedule")
-            elif 'nutrient' in indicator.lower():
-                recommendations.append("Review and adjust nutrient formulation")
-            elif 'fungal' in indicator.lower():
-                recommendations.append("Check pH and moisture for fungal health")
-            elif 'pest' in indicator.lower():
-                recommendations.append("Deploy additional beneficial insects")
-                
-        if not recommendations and overall_health > 0.9:
-            recommendations.append("System healthy - safe to proceed with monitoring")
-            
-        return recommendations
-        
-    async def get_public_summary(self) -> Dict:
-        """Get public-friendly summary of current status"""
-        current_health = self.health_history[-1] if self.health_history else None
-        
-        return {
-            'project': 'Living Quantum CEA Monitor',
-            'current_phase': self.current_phase.value,
-            'system_health': current_health.overall_health if current_health else 'Unknown',
-            'days_monitoring': len(self.baseline_measurements),
-            'natural_coherence_events': len(self.coherence_observations),
-            'next_milestone': self._get_next_milestone(),
-            'open_data_available': True,
-            'community_reviewers': len(self.community_reviewers),
-            'last_updated': datetime.utcnow().isoformat()
-        }
