--- planetary-optimizer-v3.py
@@ -1,543 +1,0 @@
-# modules/climate_adaptation/planetary_optimizer_v3.py
-"""
-OSCE Planetary Optimizer v3 - Enhanced with Production HAL Integration
-Distributed optimization with secure hardware abstraction
-
-Key v3 Enhancements:
-- Full HAL integration for all sensor operations
-- Distributed multi-site coordination
-- Hardware health-aware optimization
-- Secure federated operations
-"""
-
-import asyncio
-import json
-from typing import Dict, List, Optional, Tuple, Any
-from datetime import datetime, timedelta
-import numpy as np
-from dataclasses import dataclass
-from enum import Enum
-
-from osce.core.base import OSCEModule
-from osce.api.federation import FederatedNode
-from osce.ml.optimization import MultiObjectiveOptimizer
-from osce.utils.logging import get_logger
-from osce_hal_enhanced import HardwareManager, HardwareHealth, SensorInfo
-
-logger = get_logger(__name__)
-
-class OptimizationObjective(Enum):
-    """Optimization goals for planetary coordination"""
-    MINIMIZE_TRANSPORT = "minimize_transport"
-    MAXIMIZE_NUTRITION = "maximize_nutrition"  
-    ENSURE_FOOD_SECURITY = "ensure_food_security"
-    PRESERVE_BIODIVERSITY = "preserve_biodiversity"
-    MINIMIZE_WATER_USE = "minimize_water_use"
-    MAXIMIZE_CARBON_CAPTURE = "maximize_carbon_capture"
-    OPTIMIZE_HARDWARE_EFFICIENCY = "optimize_hardware_efficiency"  # New v3
-
-@dataclass
-class RegionalPlan:
-    """Production recommendations for a region"""
-    region_id: str
-    recommended_crops: List[Dict[str, float]]
-    water_allocation: float
-    timeline: Dict[str, datetime]
-    risk_factors: List[str]
-    adaptation_measures: List[str]
-    hardware_requirements: Dict[str, Any]  # New v3
-
-@dataclass
-class HardwareOptimizedData:
-    """Data collection optimized by hardware health"""
-    timestamp: datetime
-    data: Dict[str, Any]
-    hardware_health: Dict[str, float]
-    collection_latency_ms: float
-    sensor_reliability: float
-
-class PlanetaryOptimizerV3(OSCEModule):
-    """
-    v3: Enhanced with production HAL integration
-    
-    New Features:
-    - Hardware health-aware data collection
-    - Multi-site distributed optimization
-    - Secure sensor authentication
-    - Performance-optimized federation
-    """
-    
-    def __init__(self, config: Dict):
-        super().__init__(config)
-        self.node_id = config.get('node_id', 'osce-node-001')
-        self.region = config.get('region', 'unknown')
-        self.federation_enabled = config.get('federation_enabled', False)
-        
-        # HAL integration
-        self.hw_manager: Optional[HardwareManager] = None
-        self.sensor_groups = {}
-        self.adapter_health_threshold = config.get('adapter_health_threshold', 0.7)
-        
-        # Multi-site support
-        self.sites = config.get('sites', [{'id': 'main', 'adapters': ['local']}])
-        self.site_coordinators = {}
-        
-        # Enhanced components
-        self.federation = None
-        self.optimizer = MultiObjectiveOptimizer()
-        self.local_data_cache = {}
-        self.regional_plans = {}
-        
-        # Performance monitoring
-        self.collection_metrics = []
-        self.optimization_metrics = []
-        
-    async def initialize(self):
-        """Initialize with HAL integration"""
-        # Get hardware manager from environment
-        self.hw_manager = self.env.hal
-        
-        if not self.hw_manager:
-            logger.error("HAL not available - cannot initialize")
-            return
-            
-        # Discover and group sensors by type
-        await self._initialize_sensor_groups()
-        
-        # Setup multi-site coordinators if needed
-        if len(self.sites) > 1:
-            await self._initialize_site_coordinators()
-            
-        # Register for hardware events
-        self.hw_manager.on_event('adapter_failed', self._handle_adapter_failure)
-        self.hw_manager.on_event('hardware_discovered', self._handle_new_hardware)
-        
-        # Initialize federation if enabled
-        if self.federation_enabled:
-            await self._join_federation()
-            
-        # Start optimization loop
-        asyncio.create_task(self._optimization_loop())
-        
-        logger.info(f"Planetary Optimizer v3 initialized for region: {self.region}")
-        
-    async def _initialize_sensor_groups(self):
-        """Group sensors by type for efficient collection"""
-        all_sensors = await self.hw_manager.get_all_sensors()
-        
-        for adapter_name, sensors in all_sensors.items():
-            for sensor in sensors:
-                sensor_type = sensor.sensor_type
-                
-                if sensor_type not in self.sensor_groups:
-                    self.sensor_groups[sensor_type] = []
-                    
-                self.sensor_groups[sensor_type].append({
-                    'adapter': adapter_name,
-                    'sensor': sensor,
-                    'last_reading': None,
-                    'error_count': 0
-                })
-                
-        logger.info(f"Initialized sensor groups: {list(self.sensor_groups.keys())}")
-        
-    async def _collect_local_data(self) -> HardwareOptimizedData:
-        """Collect data with hardware health awareness"""
-        start_time = datetime.utcnow()
-        data = {
-            'timestamp': start_time.isoformat(),
-            'region': self.region,
-            'environmental': {},
-            'production': {},
-            'resources': {}
-        }
-        
-        hardware_health = {}
-        total_latency = 0
-        successful_reads = 0
-        total_reads = 0
-        
-        # Collect environmental data with health checks
-        env_data = {}
-        
-        # Temperature from best available sensor
-        temp_sensors = self.sensor_groups.get('DS18B20', []) + \
-                      self.sensor_groups.get('BME280', []) + \
-                      self.sensor_groups.get('DHT22', [])
-                      
-        if temp_sensors:
-            temp = await self._read_from_best_sensor(temp_sensors, 'temperature')
-            if temp is not None:
-                env_data['temperature'] = temp['value']
-                total_latency += temp['latency']
-                successful_reads += 1
-            total_reads += 1
-            
-        # Humidity
-        humidity_sensors = self.sensor_groups.get('BME280', []) + \
-                          self.sensor_groups.get('DHT22', [])
-                          
-        if humidity_sensors:
-            humidity = await self._read_from_best_sensor(humidity_sensors, 'humidity')
-            if humidity is not None:
-                env_data['humidity'] = humidity['value']
-                total_latency += humidity['latency']
-                successful_reads += 1
-            total_reads += 1
-            
-        # CO2
-        co2_sensors = self.sensor_groups.get('MH-Z19', []) + \
-                     self.sensor_groups.get('SCD30', [])
-                     
-        if co2_sensors:
-            co2 = await self._read_from_best_sensor(co2_sensors, 'co2')
-            if co2 is not None:
-                env_data['co2'] = co2['value']
-                total_latency += co2['latency']
-                successful_reads += 1
-            total_reads += 1
-            
-        # Light
-        light_sensors = self.sensor_groups.get('BH1750', []) + \
-                       self.sensor_groups.get('VEML7700', [])
-                       
-        if light_sensors:
-            light = await self._read_from_best_sensor(light_sensors, 'light_intensity')
-            if light is not None:
-                env_data['light_intensity'] = light['value']
-                total_latency += light['latency']
-                successful_reads += 1
-            total_reads += 1
-            
-        data['environmental'] = env_data
-        
-        # Collect hardware health metrics
-        for adapter_name, adapter in self.hw_manager.adapters.items():
-            health = await adapter.health_check()
-            hardware_health[adapter_name] = {
-                'state': health.name,
-                'score': adapter.health_score,
-                'metrics': adapter.get_metrics_summary()
-            }
-            
-        # Calculate reliability
-        sensor_reliability = successful_reads / max(total_reads, 1)
-        avg_latency = total_latency / max(successful_reads, 1)
-        
-        # Get production data
-        data['production'] = await self._get_production_data_hal()
-        
-        # Get resource data
-        data['resources'] = await self._get_resource_data_hal()
-        
-        # Cache locally
-        optimized_data = HardwareOptimizedData(
-            timestamp=datetime.utcnow(),
-            data=data,
-            hardware_health=hardware_health,
-            collection_latency_ms=avg_latency,
-            sensor_reliability=sensor_reliability
-        )
-        
-        self.local_data_cache[start_time] = optimized_data
-        
-        return optimized_data
-        
-    async def _read_from_best_sensor(self, sensor_list: List[Dict], 
-                                    reading_type: str) -> Optional[Dict[str, Any]]:
-        """Read from the healthiest available sensor"""
-        # Sort by adapter health and error count
-        sorted_sensors = sorted(
-            sensor_list,
-            key=lambda x: (
-                -self.hw_manager.adapters[x['adapter']].health_score,
-                x['error_count']
-            )
-        )
-        
-        for sensor_info in sorted_sensors:
-            adapter = self.hw_manager.get_adapter(sensor_info['adapter'])
-            
-            if not adapter or adapter.health_score < self.adapter_health_threshold:
-                continue
-                
-            try:
-                start = datetime.utcnow()
-                
-                # Secure sensor read through HAL
-                value = await adapter.secure_operation(
-                    self._read_sensor_value,
-                    sensor_info['sensor'],
-                    reading_type
-                )
-                
-                latency = (datetime.utcnow() - start).total_seconds() * 1000
-                
-                # Update last reading
-                sensor_info['last_reading'] = {
-                    'value': value,
-                    'timestamp': datetime.utcnow(),
-                    'latency': latency
-                }
-                
-                # Reset error count on success
-                sensor_info['error_count'] = 0
-                
-                return {
-                    'value': value,
-                    'latency': latency,
-                    'sensor': sensor_info['sensor'].name,
-                    'adapter': sensor_info['adapter']
-                }
-                
-            except Exception as e:
-                sensor_info['error_count'] += 1
-                logger.warning(f"Sensor read failed: {sensor_info['sensor'].name} - {e}")
-                continue
-                
-        return None
-        
-    async def _read_sensor_value(self, sensor: SensorInfo, reading_type: str) -> float:
-        """Read specific value type from sensor"""
-        # This would be implemented based on sensor type
-        # For now, return simulated value
-        if reading_type == 'temperature':
-            return 22.5 + np.random.normal(0, 0.5)
-        elif reading_type == 'humidity':
-            return 65.0 + np.random.normal(0, 2)
-        elif reading_type == 'co2':
-            return 400 + np.random.normal(0, 10)
-        elif reading_type == 'light_intensity':
-            return 500 + np.random.normal(0, 50)
-        return 0.0
-        
-    async def _get_production_data_hal(self) -> Dict:
-        """Get production data using HAL sensors"""
-        production_data = {
-            'active_crops': [],
-            'growth_stages': {},
-            'yield_projections': {},
-            'sensor_coverage': {}
-        }
-        
-        # Use camera sensors for growth monitoring
-        camera_sensors = self.sensor_groups.get('camera', [])
-        
-        for cam_info in camera_sensors:
-            adapter = self.hw_manager.get_adapter(cam_info['adapter'])
-            if adapter and adapter.health_score > 0.8:
-                try:
-                    # Capture and analyze image
-                    image_data = await adapter.secure_operation(
-                        self._capture_plant_image,
-                        cam_info['sensor']
-                    )
-                    
-                    # Analyze growth
-                    analysis = await self._analyze_plant_growth(image_data)
-                    
-                    crop_id = analysis['crop_id']
-                    production_data['active_crops'].append(crop_id)
-                    production_data['growth_stages'][crop_id] = analysis['stage']
-                    production_data['yield_projections'][crop_id] = analysis['yield_estimate']
-                    
-                except Exception as e:
-                    logger.error(f"Production monitoring failed: {e}")
-                    
-        # Calculate sensor coverage
-        total_area = await self.env.get_total_area()
-        monitored_area = len(production_data['active_crops']) * 10  # m² per sensor
-        production_data['sensor_coverage']['percentage'] = (monitored_area / total_area) * 100
-        
-        return production_data
-        
-    async def _apply_recommendations(self, plan: RegionalPlan):
-        """Apply recommendations with hardware validation"""
-        logger.info(f"Applying regional plan for {plan.region_id}")
-        
-        # Validate hardware requirements
-        hw_ready = await self._validate_hardware_requirements(plan.hardware_requirements)
-        
-        if not hw_ready:
-            logger.warning("Insufficient hardware for plan implementation")
-            return
-            
-        # Apply crop recommendations through HAL actuators
-        for crop, area in plan.recommended_crops:
-            if await self._is_suitable_crop(crop):
-                # Use HAL to control planting systems
-                await self._execute_planting_through_hal(crop, area, plan.timeline)
-                
-        # Adjust resource usage through HAL
-        await self._adjust_resources_through_hal(plan)
-        
-        # Log compliance
-        self.regional_plans[datetime.utcnow()] = plan
-        
-    async def _validate_hardware_requirements(self, requirements: Dict[str, Any]) -> bool:
-        """Check if hardware meets plan requirements"""
-        if not requirements:
-            return True
-            
-        # Check sensor availability
-        required_sensors = requirements.get('sensors', {})
-        for sensor_type, min_count in required_sensors.items():
-            available = len(self.sensor_groups.get(sensor_type, []))
-            if available < min_count:
-                logger.warning(f"Insufficient {sensor_type} sensors: {available}/{min_count}")
-                return False
-                
-        # Check actuator availability
-        required_actuators = requirements.get('actuators', {})
-        for actuator_type, min_count in required_actuators.items():
-            # Check through HAL
-            available_actuators = await self._count_available_actuators(actuator_type)
-            if available_actuators < min_count:
-                logger.warning(f"Insufficient {actuator_type} actuators")
-                return False
-                
-        # Check adapter health
-        min_health = requirements.get('min_adapter_health', 0.7)
-        for adapter in self.hw_manager.adapters.values():
-            if adapter.health_score < min_health:
-                logger.warning(f"Adapter health too low: {adapter.adapter_id}")
-                return False
-                
-        return True
-        
-    async def _execute_planting_through_hal(self, crop: str, area: float, 
-                                          timeline: Dict[str, datetime]):
-        """Execute planting using HAL-controlled systems"""
-        # Find planting actuators
-        planting_adapters = [
-            adapter for name, adapter in self.hw_manager.adapters.items()
-            if 'planting' in adapter.get_capabilities().features
-        ]
-        
-        if not planting_adapters:
-            logger.warning("No planting systems available")
-            return
-            
-        # Use healthiest adapter
-        best_adapter = max(planting_adapters, key=lambda a: a.health_score)
-        
-        # Execute planting sequence
-        await best_adapter.secure_operation(
-            self._planting_sequence,
-            crop=crop,
-            area=area,
-            start_date=timeline.get('planting_start')
-        )
-        
-    async def _handle_adapter_failure(self, event_data: Dict):
-        """Handle hardware adapter failures"""
-        failed_adapter = event_data['name']
-        logger.warning(f"Adapter failed: {failed_adapter}")
-        
-        # Find affected sensors
-        affected_sensors = []
-        for sensor_type, sensors in self.sensor_groups.items():
-            for sensor_info in sensors:
-                if sensor_info['adapter'] == failed_adapter:
-                    affected_sensors.append(sensor_info)
-                    
-        # Try to find alternative sensors
-        for sensor_info in affected_sensors:
-            alternatives = await self._find_alternative_sensors(sensor_info['sensor'])
-            
-            if alternatives:
-                logger.info(f"Found {len(alternatives)} alternatives for {sensor_info['sensor'].name}")
-                # Update sensor group with alternative
-                # ... remap sensor to alternative adapter
-                
-        # Notify federation of reduced capacity
-        if self.federation_enabled and self.federation:
-            await self.federation.update_capacity({
-                'sensor_loss': len(affected_sensors),
-                'adapter_failure': failed_adapter,
-                'timestamp': datetime.utcnow()
-            })
-            
-    async def _handle_new_hardware(self, event_data: Dict):
-        """Handle newly discovered hardware"""
-        new_devices = event_data['devices']
-        logger.info(f"New hardware discovered: {len(new_devices)} devices")
-        
-        # Re-initialize sensor groups
-        await self._initialize_sensor_groups()
-        
-        # Notify federation of increased capacity
-        if self.federation_enabled and self.federation:
-            await self.federation.update_capacity({
-                'new_devices': len(new_devices),
-                'timestamp': datetime.utcnow()
-            })
-            
-    async def multi_site_optimization(self) -> Dict[str, RegionalPlan]:
-        """Optimize across multiple physical sites"""
-        site_data = {}
-        
-        # Collect data from all sites
-        for site in self.sites:
-            site_id = site['id']
-            
-            # Get site-specific adapters
-            site_adapters = [
-                self.hw_manager.get_adapter(name) 
-                for name in site['adapters']
-            ]
-            
-            # Collect site data
-            if site_id in self.site_coordinators:
-                coordinator = self.site_coordinators[site_id]
-                data = await coordinator.collect_site_data(site_adapters)
-                site_data[site_id] = data
-                
-        # Run distributed optimization
-        optimization_result = await self.optimizer.optimize_multi_site(
-            sites=site_data,
-            objectives=[
-                OptimizationObjective.MAXIMIZE_NUTRITION,
-                OptimizationObjective.MINIMIZE_TRANSPORT,
-                OptimizationObjective.OPTIMIZE_HARDWARE_EFFICIENCY
-            ],
-            constraints={
-                'maintain_hardware_health': 0.8,
-                'sensor_coverage': 0.9
-            }
-        )
-        
-        # Generate site-specific plans
-        site_plans = {}
-        for site_id, site_optimization in optimization_result.items():
-            plan = RegionalPlan(
-                region_id=f"{self.region}_{site_id}",
-                recommended_crops=site_optimization['crops'],
-                water_allocation=site_optimization['water'],
-                timeline=site_optimization['timeline'],
-                risk_factors=site_optimization['risks'],
-                adaptation_measures=site_optimization['adaptations'],
-                hardware_requirements=site_optimization['hardware_needs']
-            )
-            site_plans[site_id] = plan
-            
-        return site_plans
-        
-    async def get_system_metrics(self) -> Dict[str, Any]:
-        """Get comprehensive system metrics including hardware"""
-        metrics = {
-            'optimization': {
-                'plans_generated': len(self.regional_plans),
-                'last_optimization': max(self.regional_plans.keys()) if self.regional_plans else None,
-                'federation_connected': self.federation is not None
-            },
-            'data_collection': {
-                'sensor_groups': len(self.sensor_groups),
-                'total_sensors': sum(len(s) for s in self.sensor_groups.values()),
-                'avg_reliability': np.mean([d.sensor_reliability for d in self.local_data_cache.values()]) if self.local_data_cache else 0,
-                'avg_latency_ms': np.mean([d.collection_latency_ms for d in self.local_data_cache.values()]) if self.local_data_cache else 0
-            },
-            'hardware': self.hw_manager.get_metrics_dashboard() if self.hw_manager else {}
-        }
-        
-        return metrics
