+++ src/osce/modules/__init__.py
@@ -1,0 +1,12 @@
+"""OSCE optional modules packaged for easy access."""
+
+from .quantum_mesh.qkd_network_v3 import QuantumSecuredMeshV3
+from .abil.plant_consciousness_interface_v3 import PlantConsciousnessInterfaceV3
+from .carbon_credits.blockchain_carbon_v3 import CarbonCreditEngineV3
+
+__all__ = [
+    "QuantumSecuredMeshV3",
+    "PlantConsciousnessInterfaceV3",
+    "CarbonCreditEngineV3",
+]
+
